# source-project/backend/src/main/resources/static/binarying-api-doc.yaml
openapi: 3.0.3
info:
  title: Binarying
  description: |-
    Mapa dos endpoints da API do projeto Binarying
    
    Projeto:
    - [Repositório](https://github.com/KawanWan/Binarying)
  version: 0.9.0
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /exercise:
    get:
        tags:
          - exercise
        summary: Retorna todos os exercícios.
        responses:
          '200':
            description: successful operation
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: '#/components/schemas/Exercise'
    post:
      tags:
        - exercise
      summary: Adiciona um novo exercicio ao banco
      description: O id não precisa ser inserido. Caso seja, há risco de colidir com outro registro no banco de mesmo id. O exercicio pode conter múltiplos conceitos separados por vírgula
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              $ref: '#/components/schemas/Exercise'
      responses:
        '201':
          description: Exercicio criado com sucesso
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem:
                    type: string
                    example: "Erro na requisição"
  /exercise/{id}:
    get:
      tags:
        - exercise
      summary: Retorna o exercicio pelo id especificado
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
    put:
      tags:
        - exercise
      summary: Atualiza um exercicio existente
      parameters:
        - name: id
          in: path
          required: true
          description: ID do exercicio a ser atualizado
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                concept:
                  type: string
                  example: "syntax, fundamentals"
                exerciseType:
                  type: string
                  example: "logical"
                difficulty:
                  type: string
                  example: "TRAINEE"
                question:
                  type: string
                  example: "O que este programa deve imprimir?"
                answers:
                  type: string
                  example: "HELLO WORLD! | Hello World! | Hello, World! | HelloWorld"
                correctAnswerIndex:
                  type: integer
                  example: 3
                codeExample:
                  type: string
                  example: 'public class HelloWorld {\n                        public static void main(String[] args) {\n                            System.out.println("Hello, World!");\n                        }\n                    }    '
                phaseId:
                  type: integer
                  example: 1
      responses:
        '201':
          description: Exercicio atualizado com sucesso
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem:
                    type: string
                    example: "Erro na requisição"
    delete:
      tags:
        - exercise
      summary: Apaga um exercicio do banco de dados pelo id
      parameters:
        - name: id
          in: path
          required: true
          description: ID do exercicio a ser apagado
          schema:
            type: integer
      responses:
        '201':
          description: Exercicio apagado com sucesso
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem:
                    type: string
                    example: "Erro na requisição"
  /exercise/random:
    get:
        tags:
          - exercise
        summary: Retorna exercícios aleatórios com os parâmetros de conceito, dificuldade e limite.
        description: O limite é utilizado para etabelecer o máximo de exercícios que deve ser retornado.
        parameters:
          - name: concept
            in: query
            required: false
            schema:
              type: string
          - name: difficulty
            in: query
            required: false
            schema:
              type: string
          - name: limit
            in: query
            required: false
            schema:
              type: integer
        responses:
          '200':
            description: successful operation
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: '#/components/schemas/Exercise'
components:
  schemas:
...
    Exercise:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        concept:
          type: string
          example: "syntax, fundamentals"
        exerciseType:
          type: string
          example: "logical"
        difficulty:
          type: string
          example: "TRAINEE"
        question:
          type: string
          example: "O que este programa deve imprimir?"
        answers:
          type: string
          example: "HELLO WORLD! | Hello World! | Hello, World! | HelloWorld"
        correctAnswerIndex:
          type: integer
          example: 3
        codeExample:
          type: string
          example: 'public class HelloWorld {\n                        public static void main(String[] args) {\n                            System.out.println("Hello, World!");\n                        }\n                    }    '
        phaseId:
          type: integer
          example: 1